Consolidate files into respective components.
