+++ src/routes/HomePage.jsx
@@ -1,36 +1,12 @@
 import "./styles/HomePage.css";
-import React, { useState, useEffect } from 'react';
-import Button from 'react-bootstrap/Button';
-import { googleLogout, useGoogleLogin } from '@react-oauth/google';
-import axios from 'axios';
+import React, { useState } from 'react';
+import GoogleLoginButton from "../components/Login/GoogleLoginButton";
+import {googleLogout} from "@react-oauth/google";
+import LoginPage from "./LoginPage";
 
-function Homepage() {
+const Homepage = () => {
     const [ user, setUser ] = useState(null);
     const [ profile, setProfile ] = useState(null);
-
-    const login = useGoogleLogin({
-        onSuccess: (codeResponse) => setUser(codeResponse),
-        onError: (error) => console.log('Login Failed:', error)
-    });
-
-    useEffect(
-        () => {
-            if (user) {
-                axios
-                    .get(`https://www.googleapis.com/oauth2/v1/userinfo?access_token=${user.access_token}`, {
-                        headers: {
-                            Authorization: `Bearer ${user.access_token}`,
-                            Accept: 'application/json'
-                        }
-                    })
-                    .then((res) => {
-                        setProfile(res.data);
-                    })
-                    .catch((err) => console.log(err));
-            }
-        },
-        [ user ]
-    );
 
     // log out function to log the user out of google and set the profile array to null
     const logOut = () => {
@@ -41,7 +17,6 @@ function Homepage() {
     return (
         <div className={"main-container"}>
             <div className={"logo-text"}>Pseudo Pass</div>
-            <div className={"login-text"}>Login using SJSU Credentials</div>
             {profile ? (
                 <div>
                     <img src={profile.picture} alt="user image" />
@@ -53,27 +28,11 @@ function Homepage() {
                     <button onClick={logOut}>Log out</button>
                 </div>
             ) : (
-
-                <Button
-                    className={"login-button"}
-                    variant="outline-dark"
-                    onClick={() => login()} >
-                    <svg version="1.1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48"
-                         className="google-icon">
-                        <g>
-                            <path className="google-icon-red" fill="#EA4335"
-                                  d="M24 9.5c3.54 0 6.71 1.22 9.21 3.6l6.85-6.85C35.9 2.38 30.47 0 24 0 14.62 0 6.51 5.38 2.56 13.22l7.98 6.19C12.43 13.72 17.74 9.5 24 9.5z"></path>
-                            <path className="google-icon-blue" fill="#4285F4"
-                                  d="M46.98 24.55c0-1.57-.15-3.09-.38-4.55H24v9.02h12.94c-.58 2.96-2.26 5.48-4.78 7.18l7.73 6c4.51-4.18 7.09-10.36 7.09-17.65z"></path>
-                            <path className="google-icon-yellow" fill="#FBBC05"
-                                  d="M10.53 28.59c-.48-1.45-.76-2.99-.76-4.59s.27-3.14.76-4.59l-7.98-6.19C.92 16.46 0 20.12 0 24c0 3.88.92 7.54 2.56 10.78l7.97-6.19z"></path>
-                            <path className="google-icon-green" fill="#34A853"
-                                  d="M24 48c6.48 0 11.93-2.13 15.89-5.81l-7.73-6c-2.15 1.45-4.92 2.3-8.16 2.3-6.26 0-11.57-4.22-13.47-9.91l-7.98 6.19C6.51 42.62 14.62 48 24 48z"></path>
-                            <path fill="none" d="M0 0h48v48H0z"></path>
-                        </g>
-                    </svg>
-                    Continue with Google
-                </Button>
+                <LoginPage
+                    user={user}
+                    setProfile={setProfile}
+                    setUser={setUser}
+                />
             )}
         </div>
     );
