+++ src/routes/styles/LoginPage.css
@@ -1,0 +1,3 @@
+.login-text {
+    margin: 24px;
+}
